change transfer ownership to a single script
